third commit

+++ database/migrations/2025_04_29_000012_create_tbl_siswa.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('tbl_siswa', function (Blueprint $table) {
-            $table->integer('idsiswa')->autoIncrement()->unsigned();
+            $table->bigInteger('idsiswa')->autoIncrement()->unsigned();
             $table->string('nis');
             $table->string('nisn');
             $table->string('namasiswa');
